Fix Timo's review comments
